+++ server/services/websocket.ts
@@ -83,6 +83,7 @@ import {
   handlePrevPhase,
   handleSetPhase,
   handleStartNextRound,
+  handleCompleteRound,
   handleStartNewMatch
 } from '../handlers/phaseManagement.js';
 
@@ -222,6 +223,7 @@ function routeMessage(ws, data) {
     'PREV_PHASE': handlePrevPhase,
     'SET_PHASE': handleSetPhase,
     'START_NEXT_ROUND': handleStartNextRound,
+    'COMPLETE_ROUND': handleCompleteRound,
     'START_NEW_MATCH': handleStartNewMatch,
     'SET_DUMMY_PLAYER_COUNT': handleSetDummyPlayerCount,
     'UPDATE_PLAYER_NAME': handleUpdatePlayerName,